refactor: split reporter options test to reduce test runtime
